// serviceLayer/src/fbs/generated/model/authenticatedPatronV8.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * FBS Adapter
 * OpenAPI spec version: 1.0
 */
import type { AuthenticatedPatronV8AuthenticateStatus } from "./authenticatedPatronV8AuthenticateStatus";
import type { PatronV7 } from "./patronV7";

export interface AuthenticatedPatronV8 {
  /** AuthenticateStatus:
 <ul>
     <li>- 'VALID': successfully authenticated</li>
     <li>- 'INVALID': either the user is not known in the system, or an invalid combination of authentication parameters has been used.</li>
     <li>- 'LOANER_LOCKED_OUT': user has been blocked temporary because of too many failed login attempts</li>
 </ul> */
  authenticateStatus: AuthenticatedPatronV8AuthenticateStatus;
  patron?: PatronV7;
}